add api connection to integrations to ping
