+++ web/ui/src/lib/query-constants.ts
@@ -21,9 +21,10 @@ export const LIST_DECKS = 'LIST_DECKS' as const;
 export const FETCH_DECK = 'FETCH_DECK' as const;
 export const FETCH_CONTACT = 'FETCH_CONTACT' as const
 export const DELETE_CONTACT = 'DELETE_CONTACT' as const
-export const UPDATE_WORKSPACE = "UPDATE_WORKSPACE" as const
-export const UPDATE_WORKSPACE_PREFERENCES = "UPDATE_WORKSPACE_PREFERENCES" as const
-export const FETCH_PREFERENCES = "FETCH_PREFERENCES" as const
-export const LIST_INTEGRATIONS = "LIST_INTEGRATIONS" as const
-export const FETCH_BILLING_PORTAL_URL = "FETCH_BILLING_PORTAL_URL" as const
-export const UPDATE_CONTACT = "UPDATE_CONTENT" as const
+export const UPDATE_WORKSPACE = "UPDATE_WORKSPACE" as const;
+export const UPDATE_WORKSPACE_PREFERENCES = "UPDATE_WORKSPACE_PREFERENCES" as const;
+export const FETCH_PREFERENCES = "FETCH_PREFERENCES" as const;
+export const LIST_INTEGRATIONS = "LIST_INTEGRATIONS" as const;
+export const FETCH_BILLING_PORTAL_URL = "FETCH_BILLING_PORTAL_URL" as const;
+export const UPDATE_CONTACT = "UPDATE_CONTENT" as const;
+export const PING_INTEGRATION = 'PING_INTEGRATION' as const;
